Create + Update Data Table Gen Files
DataTableGen.hs
- added current functions related to generating the output file DataTable.csv
- functions may be further split into a separate 'printer' script

DirectoryController.hs
- moved some functions to DataTableGen.hs and SourceCodeReader.hs
- retained iterator function; more functions related to directories/filepaths will be added later

SourceCodeReader.hs
- added current functions related to reading Haskell source code

DTG_Config.txt
- renamed old configuration/settings file (DCConfig.txt -> DTG_Config.txt)
